+++ R/arcsinh_x.R
@@ -21,8 +21,8 @@
 #' \item{x.t}{transformed 
 #'   original data} 
 #'   \item{x}{original data} 
-#'   \item{mean}{mean of vector post-transformation} 
-#'   \item{sd}{sd of vector post-transformation} 
+#'   \item{mean}{mean after transformation but prior to standardization} 
+#'   \item{sd}{sd after transformation but prior to standardization} 
 #'   \item{n}{number of nonmissing observations}
 #'   \item{norm_stat}{Pearson's P / degrees of freedom}
 #'   \item{standardize}{was the transformation standardized}
@@ -93,8 +93,8 @@ print.arcsinh_x <- function(x, ...) {
   cat(ifelse(x$standardize, "Standardized", "Non-Standardized"),
       'asinh(x) Transformation with', x$n, 'nonmissing obs.:\n', 
       'Relevant statistics:\n',
-      '- mean (post-transform) =', x$mean, '\n',
-      '- sd (post-transform) =', x$sd, '\n')
+      '- mean (before standardization) =', x$mean, '\n',
+      '- sd (before standardization) =', x$sd, '\n')
 }