+++ R/lambert.R
@@ -33,8 +33,8 @@
 #' @return A list of class \code{lambert} with elements 
 #'   \item{x.t}{transformed original data} 
 #'   \item{x}{original data} 
-#'   \item{mean}{mean of vector} 
-#'   \item{sd}{sd of vector post-BC transformation} 
+#'   \item{mean}{mean after transformation but prior to standardization} 
+#'   \item{sd}{sd after transformation but prior to standardization} 
 #'   \item{tau.mat}{estimated parameters of LambertW::Gaussianize} 
 #'   \item{n}{number of nonmissing observations}
 #'   \item{norm_stat}{Pearson's P / degrees of freedom}
@@ -152,6 +152,6 @@ print.lambert <- function(x, ...) {
       'with', x$n, 'nonmissing obs.:\n', 
       'Estimated statistics:\n',
       prettyTau,
-      '- mean (post-transform) =', x$mean, '\n',
-      '- sd (post-transform) =', x$sd, '\n')
+      '- mean (before standardization) =', x$mean, '\n',
+      '- sd (before standardization) =', x$sd, '\n')
 }