+++ NEWS.md
@@ -1,7 +1,9 @@
-# bestNormalize 1.2.0.9000
+# bestNormalize 1.2.0.9001
 
+- Add "no_transform" function - does the same thing as I(x) but in the syntax of other transformations 
+  (this allows the normalization statistics to also be calculated if no transformation is performed). 
 - Add support for lambert transforms of type "h" in the `bestNormalize` function via `allow_lambert_h` argument.
-- Add "post-transformation" to printout of different transforms' means and sds to clarify output
+- Add "before standardization" to printout of different transforms' means and sds to clarify output
 
 # bestNormalize 1.2.0